+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+
+- Added ColumnDropper that allows a user to specify the columns that should not be extracted in the pipeline.
 - Added Favourite/UnFavourite to right click context menus
 
 ### Fixed